Move remaining code out of the store directory
