+++ src/views/application-container.tsx
@@ -1,7 +1,7 @@
 import {h} from 'preact'
-import {getCanvasContext} from '../store/getters'
+import {getCanvasContext} from '../app-state/getters'
 import {memo, useMemo} from 'preact/compat'
-import {useActiveProfileState} from '../store'
+import {useActiveProfileState} from '../app-state/active-profile-state'
 import {useTheme} from './themes/theme'
 import {
   dragActiveAtom,
