+++ src/views/search-view.tsx
@@ -5,7 +5,7 @@ import {memo} from 'preact/compat'
 import {Sizes, FontSize} from './style'
 import {ProfileSearchResults} from '../lib/profile-search'
 import {Profile} from '../lib/profile'
-import {useActiveProfileState} from '../store'
+import {useActiveProfileState} from '../app-state/active-profile-state'
 import {useTheme, withTheme} from './themes/theme'
 import {searchIsActiveAtom, searchQueryAtom} from '../app-state'
 import {useAtom} from '../lib/atom'
